Allow target origin for oauth success message to be specified
The `postMessage` from the OAuth window to the opening window specifies an origin for the target window. Currently this is set to the same as the OAuth window which will not work locally or in situations where the DNS has not been set up to allow the auth function to exist within the primary application's hostname.

This (somewhat hacky) change allows users to specify an alternate `TARGET_ORIGIN` to allow the `postMessage` to complete successfully without resorting to allowing calls from any origin.